requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
python-dotenv
Flask-Migrate
